Splits MedScanner.tsx into files for constants, data and helper(s) (goodbye MedScanner commit history you lasted 8-9 months before I renamed you again)
