+++ tgui/packages/tgui/interfaces/MedScanner/constants.ts
@@ -1,0 +1,26 @@
+// Sizing
+/** Font size of stuff like the OD limit or organ max health */
+export const RESERVE_FONT_SIZE = '75%';
+/** What most elements will use for their padding and margin */
+export const SPACING_PIXELS = '5px';
+
+// Colors
+/** For the multiple elements using zebra stripes, this is the base for each step */
+export const ZEBRA_BG_COLOR = 'hsla(0, 0%, 100%, 0.075)';
+/** Color of the [Robotic] tag, and robotic limb names */
+export const ROBOT_LIMB_COLOR = 'hsl(218, 60%, 72%)';
+/** Custom color for brute damage */
+export const COLOR_BRUTE = 'red';
+/** Custom color for burn damage */
+export const COLOR_BURN = 'hsl(39, 100%, 60%)';
+/** Middle ground between grey and light grey color presets */
+export const MID_GREY = 'hsl(0, 0%, 59%)';
+/** Marginally darker version of the red color preset */
+export const DARKER_RED = 'hsl(0, 72%, 42%)';
+/** Marginally darker version of the orange color preset */
+export const DARKER_ORANGE = 'hsl(24, 89%, 40%)';
+/** Saturation and Luminance for `getLimbColor` */
+export const LIMB_DAMAGE_HSL = {
+  sat: '100%',
+  lum: '62%',
+};
